ExemplosJDBC/src/java/main/davi/domain/Cliente.java
/**
 * 
 */
package davi.domain;

/**
 * @author Davi
 *
 */
public class Cliente {
	
	private Long id;
	
	private String nome;
	
	private Integer idade;
	
	private String sexo;
	
	private int cpf;

	public Long getId() {
		return id;
	}

	public void setId(Long id) {
		this.id = id;
	}

	public String getNome() {
		return nome;
	}

	public void setNome(String nome) {
		this.nome = nome;
	}

	public Integer getIdade() {
		return idade;
	}

	public void setIdade(Integer idade) {
		this.idade = idade;
	}

	public String getSexo() {
		return sexo;
	}

	public void setSexo(String sexo) {
		this.sexo = sexo;
	}

	public int getCpf() {
		return cpf;
	}

	public void setCpf(int cpf) {
		this.cpf = cpf;
	}
}
